Revert "Update"
This reverts commit bebaa78521a66b31e5472b42f5c6ca8c5af19fb5.

+++ Sources/ALSA/ALSAAudio.swift
@@ -189,9 +189,9 @@ public final class ALSACaptureStream: @unchecked Sendable {
     }
 
     deinit {
-        process.terminate()
-        pipe.standardOutput.close()
-        pipe.standardError.close()
+        if process.isRunning {
+            process.terminate()
+        }
     }
 
     /// Read audio frames into a buffer
@@ -228,7 +228,9 @@ public final class ALSACaptureStream: @unchecked Sendable {
 
     /// Stop the capture
     public func stop() {
-        process.terminate()
+        if process.isRunning {
+            process.terminate()
+        }
     }
 }
